+++ system/configuration.nix
@@ -108,9 +108,15 @@ with lib;
   # Start controller
   systemd.services.playos-controller = {
     description = "PlayOS Controller";
-    serviceConfig.ExecStart = "${pkgs.playos-controller}/bin/playos-controller ${config.playos.updateUrl}";
-    serviceConfig.User = "root";
+    serviceConfig = {
+      ExecStart = "${pkgs.playos-controller}/bin/playos-controller ${config.playos.updateUrl}";
+      User = "root";
+      RestartSec = "10s";
+      Restart = "always";
+    };
     wantedBy = [ "multi-user.target" ];
+    requires = [ "rauc" "connman" ];
+    after = [ "rauc" "connman" ];
   };
 
 }